requirements.txt
pytest>=8.2,<9.0
playwright>=1.45,<2.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
